refactor(ways): remove BM25 from CLI surface (ADR-125)
Finish the cleanup started in the previous commit:

- config.rs: drop bm25_threshold, matcher, parent_threshold_multiplier
  fields (no tier to configure, no parent-threshold lowering since
  thresholds are per-node only); remove template comments that named
  them; update unit tests
- agents/mod.rs: delete bm25_stemmer_for and is_bm25_available (no
  stemmer selection, no availability check with embedding required)
- cmd/status.rs: engine has one option ("embedding" or "none"); drop
  the "configured" matcher display
- cmd/language.rs: drop BM25-fallback warnings; missing model is a
  hard not-matching condition, not a degraded mode
- cmd/template.rs + main.rs: drop --threshold from `ways template`
  (it scaffolded BM25 threshold, no longer meaningful)
- cmd/corpus.rs: tip text now says "embedding engine required" instead
  of "98% vs 91% BM25"
- cmd/render.rs: remove "semantic:bm25" trigger format
- main.rs: drop "BM25 fallback" from `ways match` help, remove
  "semantic:bm25" from scan trigger channel help
- Rename cmd/match_bm25.rs -> cmd/match_cmd.rs (the file was already
  rewritten to embedding-only in the previous commit; this finishes
  the rename)

cargo check clean. The only remaining "BM25" string in src/ is a
single comment explaining why the threshold field default changed —
that's intentional archaeology.
