+++ tools/ways-cli/src/cmd/embed.rs
@@ -2,5 +2,5 @@ use anyhow::Result;
 
 /// `ways embed` — retained as an alias for `ways match` (embedding-only since ADR-125).
 pub fn run(query: String, corpus: Option<String>, _model: Option<String>) -> Result<()> {
-    super::match_bm25::run(query, corpus)
+    super::match_cmd::run(query, corpus)
 }
