+++ tools/ways-cli/src/cmd/match_cmd.rs
