+++ tools/ways-cli/src/config.rs
@@ -25,29 +25,20 @@ pub fn global() -> &'static Config {
 /// Ways configuration.
 #[derive(Debug, Clone)]
 pub struct Config {
-    /// Default BM25 threshold for ways without explicit threshold
-    pub bm25_threshold: f64,
-    /// Parent threshold multiplier (child threshold = parent * this)
-    pub parent_threshold_multiplier: f64,
     /// Default scope for ways without explicit scope
     pub default_scope: String,
     /// Output language (e.g., "en", "ja", "auto")
     pub language: String,
     /// Disabled domains (e.g., ["ea", "itops"])
     pub disabled_domains: Vec<String>,
-    /// Matcher preference: "auto", "embedding", "bm25"
-    pub matcher: String,
 }
 
 impl Default for Config {
     fn default() -> Self {
         Self {
-            bm25_threshold: 2.0,
-            parent_threshold_multiplier: 0.8,
             default_scope: "agent".to_string(),
             language: "auto".to_string(),
             disabled_domains: Vec::new(),
-            matcher: "auto".to_string(),
         }
     }
 }
@@ -95,14 +86,6 @@ impl Config {
                 .filter_map(|v| v.as_str().map(|s| s.to_string()))
                 .collect();
         }
-
-        // Accept both field names: "engine" (new) and "semantic_engine" (legacy)
-        if let Some(engine) = v.get("engine")
-            .or_else(|| v.get("semantic_engine"))
-            .and_then(|v| v.as_str())
-        {
-            self.matcher = engine.to_string();
-        }
     }
 
     /// Apply values from a YAML config file.
@@ -117,15 +100,6 @@ impl Config {
 
         if let Some(v) = doc.get("language").and_then(|v| v.as_str()) {
             self.language = v.to_string();
-        }
-        if let Some(v) = doc.get("matcher").and_then(|v| v.as_str()) {
-            self.matcher = v.to_string();
-        }
-        if let Some(v) = doc.get("bm25_threshold").and_then(|v| v.as_f64()) {
-            self.bm25_threshold = v;
-        }
-        if let Some(v) = doc.get("parent_threshold_multiplier").and_then(|v| v.as_f64()) {
-            self.parent_threshold_multiplier = v;
         }
         if let Some(v) = doc.get("default_scope").and_then(|v| v.as_str()) {
             self.default_scope = v.to_string();
@@ -152,8 +126,6 @@ impl Config {
 # Project scope: {project}/.claude/ways.yaml (layered on top)
 
 # language: en          # Output language (en, ja, auto)
-# matcher: auto         # Matching engine (auto, embedding, bm25)
-# bm25_threshold: 2.0   # Default BM25 score threshold
 # default_scope: agent  # Default scope for ways without explicit scope
 # disabled_domains: []  # Domains to disable (e.g., [ea, itops])
 ";
@@ -187,28 +159,23 @@ mod tests {
     #[test]
     fn default_values() {
         let cfg = Config::default();
-        assert_eq!(cfg.bm25_threshold, 2.0);
         assert_eq!(cfg.language, "auto");
-        assert_eq!(cfg.matcher, "auto");
         assert_eq!(cfg.default_scope, "agent");
     }
 
     #[test]
     fn apply_yaml_overrides() {
         let mut cfg = Config::default();
-        cfg.apply_yaml("language: ja\nbm25_threshold: 1.5\nmatcher: embedding");
+        cfg.apply_yaml("language: ja");
         assert_eq!(cfg.language, "ja");
-        assert_eq!(cfg.bm25_threshold, 1.5);
-        assert_eq!(cfg.matcher, "embedding");
     }
 
     #[test]
     fn apply_ways_json() {
         let mut cfg = Config::default();
-        cfg.apply_ways_json(r#"{"output_language":"de","disabled":["ea"],"engine":"bm25"}"#);
+        cfg.apply_ways_json(r#"{"output_language":"de","disabled":["ea"]}"#);
         assert_eq!(cfg.language, "de");
         assert_eq!(cfg.disabled_domains, vec!["ea"]);
-        assert_eq!(cfg.matcher, "bm25");
     }
 
     #[test]
